Compose authors changed to save missing authors
Save missing authors will find authors from GitHub.

setCoauthors from the core will manage the work to set selected co authors and throw errors.
